+++ packages/components-angular/src/app/pages/modal-basic.component.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
 
 @Component({
   selector: 'page-modal-basic',
@@ -19,5 +19,6 @@ import { Component } from '@angular/core';
       <p-modal [heading]="'Some Heading'" [open]="true">Some Content</p-modal>
     </div>
   `,
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class ModalBasicComponent {}