+++ packages/components/src/components/navigation/link/link.tsx
@@ -24,10 +24,10 @@ export class Link {
   @Prop() public icon?: IconName = 'arrow-head-right';
 
   /** A custom URL path to a custom icon. */
-  @Prop() public iconSource?: string = undefined;
+  @Prop() public iconSource?: string;
 
   /** When providing an url then the component will be rendered as `<a>`. */
-  @Prop() public href?: string = undefined;
+  @Prop() public href?: string;
 
   /** Adapts the link color when used on dark background. */
   @Prop() public theme?: Theme = 'light';
@@ -36,10 +36,10 @@ export class Link {
   @Prop() public target?: LinkTarget = '_self';
 
   /** Special download attribute to open native browser download dialog if target url points to a downloadable file. */
-  @Prop() public download?: string = undefined;
+  @Prop() public download?: string;
 
   /** Specifies the relationship of the target object to the link object. */
-  @Prop() public rel?: string = undefined;
+  @Prop() public rel?: string;
 
   /** Show or hide label. For better accessibility it is recommended to show the label. */
   @Prop() public hideLabel?: BreakpointCustomizable<boolean> = false;