+++ packages/components/src/utils/index.ts
@@ -11,6 +11,7 @@ export * from './line-height';
 export * from './mutation-observer';
 export * from './prefix';
 export * from './property-observer';
+export * from './scrolling';
 export * from './slotted-styles';
 export * from './ssr-handling';
 export * from './styles';